+++ packages/ng-async-injector/src/lib/providers/provide-async-injector.function.ts
@@ -1,8 +1,7 @@
 import { StaticClassProvider } from '@angular/core';
 
 import { AsyncInjector } from '../injector/async-injector';
-import { R3AsyncInjector } from '../injector/r3-async-injector';
 
 export function provideAsyncInjector(): StaticClassProvider {
-  return { provide: AsyncInjector, useClass: R3AsyncInjector, deps: [] };
+  return { provide: AsyncInjector, useClass: AsyncInjector, deps: [] };
 }
